Simplify enrollment reconciliation
